move createWindow to windowsReducer fix #9962
